+++ components/pages/common/RegistrationConfirmBtn.vue
