+++ components/pages/common/RegistrationMenu.vue
