+++ components/pages/common/RegistrationRequest.vue
