+++ components/pages/common/RegistrationUserInfo.vue
